requirements.txt
alembic==1.16.5
beautifulsoup4==4.14.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deep-translator==1.11.4
filelock==3.20.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
fsspec==2025.12.0
greenlet==3.3.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
msgspec==0.20.0
networkx==3.6.1
numpy==2.2.2
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20251107
pillow==12.0.0
pycparser==2.23
PyMuPDF==1.26.6
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.4
